refactor: extract owner to its own module
